requirements.txt
pytest                   # Pytest is a testing framework which allows us to write test codes using python
pydantic
google-genai